# docs/openapi.yaml
openapi: 3.1.3
info:
  title: Evently API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /v1/health:
    get:
      responses:
        '200': { description: OK }
  /v1/events:
    get:
      parameters:
        - in: query
          name: limit
          schema: { type: integer }
        - in: query
          name: offset
          schema: { type: integer }
      responses:
        '200': { description: OK }
  /v1/events/{id}:
    get:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /v1/events/{id}/book:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user_id: { type: string }
                seats: { type: array, items: { type: string } }
                idempotency_key: { type: string }
      responses:
        '202': { description: Accepted }
        '200': { description: OK }
        '409': { description: Conflict }
  /v1/bookings/{id}/cancel:
    post:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /v1/users/{id}/bookings:
    get:
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200': { description: OK }
  /v1/waitlist/{event_id}/join:
    post:
      responses:
        '200': { description: OK }
  /v1/waitlist/{event_id}/{user_id}/optout:
    post:
      responses:
        '200': { description: OK }
  /admin/events:
    post:
      responses:
        '201': { description: Created }
  /admin/analytics:
    get:
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date-time }
        - in: query
          name: to
          schema: { type: string, format: date-time }
      responses:
        '200': { description: OK }